Adding tab-closed event
adding a 'tab-closed' event to match the 'content' event and then renamed the 'content' event to 'tab-opening'.

Now when a user changes tabs on the sidebar the tab closing event fires letting listeners know which tab is closing, before the dom fires the 'tab-opening' event to let users know which tab is opening.  This is useful for adding and removing event listeners which are attached to a particular tab.